Remove IDocumentationReferenceResolver
