+++ src/BitzArt.XDoc/DocumentationReferenceResolver.cs
@@ -9,7 +9,7 @@ namespace BitzArt.XDoc;
 /// This interface is used to extract documentation references from XML documentation comments,
 /// such as cref links and inheritdoc elements, to create structured representation of code references.
 /// </remarks>
-public interface IDocumentationReferenceResolver
+public class DocumentationReferenceResolver
 {
     /// <summary>
     /// Extracts a documentation reference from the provided XML node.
@@ -20,5 +20,19 @@ public interface IDocumentationReferenceResolver
     /// A <see cref="DocumentationReference"/> object if a reference can be extracted;
     /// otherwise, <see langword="null"/>.
     /// </returns>
-    DocumentationReference? GetReference(XDoc source, XmlNode node);
+    public DocumentationReference? GetReference(XDoc source, XmlNode node)
+    {
+        var cref = node.Attributes?["cref"]?.Value ?? string.Empty;
+
+        var isCref = !string.IsNullOrWhiteSpace(cref);
+        var isInheritDoc = node.Name == "inheritdoc";
+        var isSee = node.Name == "see";
+
+        if (!isInheritDoc && !isCref && !isSee)
+        {
+            return null;
+        }
+
+        return new DocumentationReference(node, null, cref);
+    }
 }
