--- src/BitzArt.XDoc/Resolvers/SimpleDocumentationReferenceResolver.cs
@@ -1,24 +1,0 @@
-using System.Xml;
-
-namespace BitzArt.XDoc;
-
-/// <inheritdoc />
-public class SimpleDocumentationReferenceResolver : IDocumentationReferenceResolver
-{
-    /// <inheritdoc />
-    public DocumentationReference? GetReference(XDoc source, XmlNode node)
-    {
-        var cref = node.Attributes?["cref"]?.Value ?? string.Empty;
-
-        var isCref = !string.IsNullOrWhiteSpace(cref);
-        var isInheritDoc = node.Name == "inheritdoc";
-        var isSee = node.Name == "see";
-
-        if (!isInheritDoc && !isCref && !isSee)
-        {
-            return null;
-        }
-
-        return new DocumentationReference(node, null, cref);
-    }
-}
