+++ tests/Render/BitzArt.XDoc.PlainText.Tests/TestPropertyDocumentation.cs
@@ -5,7 +5,7 @@ namespace BitzArt.XDoc.Tests;
 public class TestPropertyDocumentation : MemberDocumentation
 {
     public TestPropertyDocumentation(string xml)
-        : base(new XDoc(new SimpleDocumentationReferenceResolver()), GetXmlNode(xml))
+        : base(new XDoc(new DocumentationReferenceResolver()), GetXmlNode(xml))
     {
     }
